+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed:
+- Renamed `Dynamite` class to `ItemManager`
+- Renamed `DynamiteRegistry` to `ItemManagerRegistry`
+
 ## [0.0.3] - 2021-03-13
 ### Added:
 - Changing case of params injected to keys in `@DuplicateTo`